Revert "mm, vmpressure: int cast vmpressure level/model for -1 comparison"
This reverts commit 4462bcd3dcf1db0cc10a5c31ba1cafeff69004b3.

+++ mm/vmpressure.c
@@ -400,7 +400,7 @@ int vmpressure_register_event(struct mem_cgroup *memcg,
 	/* Find required level */
 	token = strsep(&spec, ",");
 	level = str_to_level(token);
-	if ((int)level == -1) {
+	if (level == -1) {
 		ret = -EINVAL;
 		goto out;
 	}
@@ -409,7 +409,7 @@ int vmpressure_register_event(struct mem_cgroup *memcg,
 	token = strsep(&spec, ",");
 	if (token) {
 		mode = str_to_mode(token);
-		if ((int)mode == -1) {
+		if (mode == -1) {
 			ret = -EINVAL;
 			goto out;
 		}
